fix: the type of the `orderValue` variable was changed from `double` to `BigDecimal`

+++ problema02/src/main/java/br/furb/problema02/factories/OrderTypeFactory.java
@@ -5,9 +5,11 @@ import br.furb.problema02.order.IOrderType;
 import br.furb.problema02.order.OrderTypeBuy;
 import br.furb.problema02.order.OrderTypeSell;
 
+import java.math.BigDecimal;
+
 public class OrderTypeFactory {
-	public static IOrderType CreateOrder(String investorName, double orderValue, OrderTypeEnum orderType) throws IllegalArgumentException{
-		if(orderType == null) {
+	public static IOrderType CreateOrder(String investorName, BigDecimal orderValue, OrderTypeEnum orderType) throws IllegalArgumentException {
+		if (orderType == null) {
 			throw new IllegalArgumentException("Tipo de ordem inválido!");
 		}
