+++ problema02/src/main/java/br/furb/problema02/order/IOrderType.java
@@ -2,8 +2,10 @@ package br.furb.problema02.order;
 
 import br.furb.problema02.enums.OrderTypeEnum;
 
+import java.math.BigDecimal;
+
 public interface IOrderType {
 	String getInvestorName();
-	double getOrderValue();
+	BigDecimal getOrderValue();
 	OrderTypeEnum getOrderType();
 }
