+++ problema02/src/main/java/br/furb/problema02/order/OrderTypeBuy.java
@@ -3,8 +3,10 @@ package br.furb.problema02.order;
 import br.furb.problema02.enums.OrderTypeEnum;
 import br.furb.problema02.model.Order;
 
+import java.math.BigDecimal;
+
 public class OrderTypeBuy extends Order{
-	public OrderTypeBuy(String investorName, double orderValue) {
+	public OrderTypeBuy(String investorName, BigDecimal orderValue) {
 		super(investorName, orderValue);
 	}
